Final Integration

+++ src/main/java/com/project/apprentice/repos/FacultyUserRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.project.apprentice.model.Faculty;
 
-public interface FacultyUserRepository extends JpaRepository <Faculty, Long> 
+public interface FacultyUserRepository extends JpaRepository <Faculty, Integer> 
 {
 	public Faculty findByUserIdAndPassword(int userId, String password);
 }
