+++ src/main/java/com/project/apprentice/repos/SchoolYearRepository.java
@@ -1,8 +1,9 @@
 package com.project.apprentice.repos;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import com.project.apprentice.model.SchoolYear;
 
-public interface SchoolYearRepository extends JpaRepository <SchoolYear, Long> {
-
+public interface SchoolYearRepository extends JpaRepository <SchoolYear, Integer>{
 }
+
